Move path to a separate file
